add definitions folder in api db models and migrate code structure
